storage: refactor file cache code for coming fscache
Move code to manage cached file object from filecache/cache_entry.rs
into cachedfile.rs, so it could shared by fscache subsystem later.
